ReflectionUtils extracted.
